+++ scripts/components/result/scorelist.lua
@@ -402,6 +402,10 @@ local ScoreList = {
   ---@param dt deltaTime
   handleChange = function(this, dt)
     if (this.state.scoreCount > 1) then
+      if (this.currScore ~= this.state.currScore) then
+        this.currScore = this.state.currScore;
+      end
+
       if ((not this.pressedFXL) and pressed('FXL')) then
         if ((this.currScore - 1) < 1) then
           this.currScore = this.state.scoreCount;
@@ -420,6 +424,8 @@ local ScoreList = {
 
       this.pressedFXL = pressed('FXL');
       this.pressedFXR = pressed('FXR');
+
+      this.state.currScore = this.currScore;
     end
 
     this.list:handleChange(dt, { watch = this.currScore });